requirements.txt
aiosmtplib @ file:///home/conda/feedstock_root/build_artifacts/aiosmtplib_1735226506650/work
aiosqlite==0.21.0
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1733247046149/work
anyio @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_anyio_1742243108/work
argon2-cffi @ file:///home/conda/feedstock_root/build_artifacts/argon2-cffi_1733311059102/work
argon2-cffi-bindings @ file:///D:/bld/argon2-cffi-bindings_1725356678807/work
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1733235340728/work
asyncpg @ file:///D:/bld/asyncpg_1729845844183/work
bcrypt==4.0.1
blinker @ file:///C:/b/abs_b1i87khtob/croot/blinker_1737448732095/work
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
CacheControl==0.14.3
cachetools==5.5.2
certifi @ file:///C:/b/abs_67mh6i36ci/croot/certifi_1752653721535/work/certifi
cffi @ file:///D:/bld/cffi_1725560655177/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click @ file:///D:/bld/click_1747811308011/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
cryptography @ file:///D:/bld/cryptography-split_1747571564667/work
distro==1.9.0
dnspython @ file:///C:/b/abs_449_z2xz29/croot/dnspython_1703097012342/work
email_validator @ file:///home/conda/feedstock_root/build_artifacts/email-validator-meta_1733300719943/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
fastapi @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_fastapi_1742822207/work
fastapi-cli @ file:///home/conda/feedstock_root/build_artifacts/fastapi-cli_1734302308128/work
fastapi-mail @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_fastapi-mail_1732387602/work
firebase_admin==7.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet @ file:///C:/b/abs_bbl_18a2tm/croot/greenlet_1733860081977/work
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httplib2==0.22.0
httptools @ file:///D:/bld/httptools_1732707713283/work
httpx==0.28.1
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
iniconfig @ file:///home/linux1/recipes/ci/iniconfig_1610983019677/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
jiter==0.10.0
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1733250460757/work
MarkupSafe @ file:///D:/bld/markupsafe_1733219601796/work
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1733255585584/work
msgpack==1.1.1
openai==1.97.1
packaging @ file:///C:/b/abs_3by6s2fa66/croot/packaging_1734472138782/work
passlib==1.7.4
pluggy @ file:///C:/b/abs_dfec_m79vo/croot/pluggy_1733170145382/work
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic @ file:///home/conda/feedstock_root/build_artifacts/pydantic_1746631911634/work
pydantic-settings @ file:///C:/b/abs_fe1okp6jbe/croot/pydantic-settings_1730800848545/work
pydantic_core @ file:///D:/bld/bld/rattler-build_pydantic-core_1746625389/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1732782409051/work
pyngrok==7.2.12
pyparsing==3.2.3
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
pytest @ file:///C:/b/abs_f8ndk_y_i9/croot/pytest_1750679161119/work
pytest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pytest-asyncio_1748305379/work/dist
pytest-mock==3.14.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dotenv_1750789290/work
python-multipart @ file:///home/conda/feedstock_root/build_artifacts/python-multipart_1734420773152/work
PyYAML @ file:///D:/bld/pyyaml_1737454663129/work
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
rfc3986 @ file:///Users/ktietz/demo/mc3/conda-bld/rfc3986_1629478296451/work
rich @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rich_1743371105/work/dist
rich-toolkit @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rich-toolkit_1733750834/work
rsa==4.9.1
setuptools==78.1.1
shellingham @ file:///home/conda/feedstock_root/build_artifacts/shellingham_1733300899265/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1733244044561/work
SQLAlchemy @ file:///D:/bld/sqlalchemy_1747298855422/work
starlette @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_starlette_1744661396/work
tqdm==4.67.1
typer==0.15.3
typer-slim==0.15.3
typing-inspection @ file:///home/conda/feedstock_root/build_artifacts/typing-inspection_1747870647094/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1744302253/work
uritemplate==4.2.0
urllib3 @ file:///C:/b/abs_7bst06lizn/croot/urllib3_1737133657081/work
uvicorn @ file:///D:/bld/uvicorn_1745173471370/work
watchfiles @ file:///D:/bld/watchfiles_1744451374857/work
websockets @ file:///D:/bld/websockets_1741285486349/work
wheel==0.45.1
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
